allow environment vars TRY_PATH and TRY_CONFIG to override the defaults
TRY_CONFIG specifies either a filename inside the config directory or
a path to a config file. The extension is assumed to be .toml if not
specified. This config file will be used in place of the default
config file.

TRY_PATH specifies the path where your experiments will be
stored. Always overrides the path specified in the config file.